Migrate Media type to BaseItemDto from jellyfin_api

+++ src/utils/ticks.rs
@@ -1,5 +1,5 @@
-pub fn ticks_to_seconds(ticks: usize) -> usize {
-    ticks / 10_000_000
+pub fn ticks_to_seconds(ticks: i64) -> isize {
+    ticks as isize / 10_000_000
 }
 
 pub fn seconds_to_ticks(seconds: usize) -> usize {
